Fix bug which rerendered app forever
There was a bug in useActivitySearch related to using a temporary empty array
for state. This empty array was being used as a dependency to useEffect, so the component
would rerender every time it rendered because the array was being set to a new array each
render. In other words, an infinite loop.