Move errors to temp folder
For whatever reason, the `server/src/repositories/errors` directory was checked into git
as `server/src/repositories/Errors`. Things would run fine locally, but once deployed to
Heroku, the inconsistency caused the app to crash (due to Linux having case-sensitive
paths).

In order to fix this issue, this commit will rename the directory to "temp", and the next
will change it back to "errors" (lowercase, this time).