Make jest rootDir server
When the `npm run test` command is run in the root directory or in the
server directory, we want to only run the tests in the server directory.

This is mainly because the tests in the frontend must be run with react-scripts,
and will throw errors if they are run with the default jest configuration.